Factor out table component and add performant checkbox
