+++ client/modules/datafiles/src/FileListing/FileListingTable/FileListingTable.module.css
@@ -1,0 +1,122 @@
+.listing-table-base {
+  height: 100%;
+}
+
+.checkbox-wrapper {
+  max-width: 100%;
+  font-weight: bold;
+  margin: 0;
+  padding: 0;
+  color: rgba(0, 0, 0, 0.88);
+  font-size: 14px;
+  line-height: 1.5714285714285714;
+  list-style: none;
+  display: inline-flex;
+  align-items: baseline;
+  cursor: pointer;
+  box-sizing: border-box;
+}
+.checkbox-wrapper::after {
+  display: inline-block;
+  width: 0;
+  overflow: hidden;
+  content: '\a0';
+}
+
+.checkbox {
+  font-weight: bold;
+  margin: 0;
+  padding: 0;
+  color: rgba(0, 0, 0, 0.88);
+  font-size: 14px;
+  line-height: 1;
+  list-style: none;
+  position: relative;
+  white-space: nowrap;
+  cursor: pointer;
+  border-radius: 0;
+  align-self: center;
+  box-sizing: border-box;
+}
+
+.checkbox-inner {
+  font-weight: bold;
+  color: rgba(0, 0, 0, 0.88);
+  font-size: 14px;
+  line-height: 1;
+  list-style: none;
+  white-space: nowrap;
+  cursor: pointer;
+  box-sizing: border-box;
+  display: block;
+  width: 16px;
+  height: 16px;
+  direction: ltr;
+  background-color: #ffffff;
+  border: 1px solid #d9d9d9;
+  border-radius: 0;
+  border-collapse: separate;
+  transition: all 0.3s;
+}
+
+.checkbox-inner::after {
+  box-sizing: border-box;
+  position: absolute;
+  top: 50%;
+  inset-inline-start: 25%;
+  display: table;
+  width: 5.7142857142857135px;
+  height: 9.142857142857142px;
+  border: 2px solid #fff;
+  border-top: 0;
+  border-inline-start: 0;
+  transform: rotate(45deg) scale(0) translate(-50%, -50%);
+  opacity: 0;
+  content: '';
+  transition: all 0.1s cubic-bezier(0.71, -0.46, 0.88, 0.6), opacity 0.1s;
+}
+
+.checkbox-input {
+  list-style: none;
+  white-space: nowrap;
+  color: inherit;
+  vertical-align: baseline;
+  font-family: Roboto, 'Helvetica Neue', sans-serif;
+  font-size: 100%;
+  padding: 0;
+  line-height: normal;
+  box-sizing: border-box;
+  position: absolute;
+  inset: 0;
+  z-index: 1;
+  cursor: pointer;
+  opacity: 0;
+  margin: 0;
+}
+
+.checkbox-wrapper-checked {
+  max-width: 100%;
+  font-weight: bold;
+  margin: 0;
+  padding: 0;
+  color: rgba(0, 0, 0, 0.88);
+  font-size: 14px;
+  line-height: 1.5714285714285714;
+  list-style: none;
+  display: inline-flex;
+  align-items: baseline;
+  cursor: pointer;
+  box-sizing: border-box;
+  margin-inline-start: 0;
+}
+
+.checkbox-checked .checkbox-inner::after {
+  opacity: 1;
+  transform: rotate(45deg) scale(1) translate(-50%, -50%);
+  transition: all 0.2s cubic-bezier(0.12, 0.4, 0.29, 1.46) 0.1s;
+}
+
+.checkbox-checked .checkbox-inner {
+  background-color: #337ab7;
+  border-color: #337ab7;
+}
